Update namespace organization within the Json rule engine.  Moved JsonPathResolver to Utilities project (and its interface to Types).
