+++ src/Analyzer.JsonRuleEngine.FunctionalTests/Analyzer.JsonRuleEngine.FunctionalTests.csproj
@@ -5,7 +5,7 @@
 
     <IsPackable>false</IsPackable>
 
-    <AssemblyName>Microsoft.Azure.Templates.Analyzer.JsonRuleEngine.FunctionalTests</AssemblyName>
+    <AssemblyName>Microsoft.Azure.Templates.Analyzer.RuleEngines.JsonEngine.FunctionalTests</AssemblyName>
 
     <RootNamespace>Microsoft.Azure.Templates.Analyzer.JsonRuleEngine.FunctionalTests</RootNamespace>
   </PropertyGroup>
