+++ src/Analyzer.JsonRuleEngine/InternalsVisibleTo.cs
@@ -3,8 +3,8 @@
 
 using System.Runtime.CompilerServices;
 
-[assembly: InternalsVisibleTo("Microsoft.Azure.Templates.Analyzer.JsonRuleEngine.UnitTests")]
-[assembly: InternalsVisibleTo("Microsoft.Azure.Templates.Analyzer.JsonRuleEngine.FunctionalTests")]
+[assembly: InternalsVisibleTo("Microsoft.Azure.Templates.Analyzer.RuleEngines.JsonEngine.UnitTests")]
+[assembly: InternalsVisibleTo("Microsoft.Azure.Templates.Analyzer.RuleEngines.JsonEngine.FunctionalTests")]
 
 // For Moq testing library
 [assembly: InternalsVisibleTo("DynamicProxyGenAssembly2")]
