+++ src/Analyzer.JsonRuleEngine/Operators/EqualsOperator.cs
@@ -3,7 +3,7 @@
 
 using Newtonsoft.Json.Linq;
 
-namespace Microsoft.Azure.Templates.Analyzer.JsonRuleEngine
+namespace Microsoft.Azure.Templates.Analyzer.RuleEngines.JsonEngine.Operators
 {
     internal class EqualsOperator : LeafExpressionOperator
     {
