+++ src/Analyzer.Utilities.UnitTests/JsonPathResolverTests.cs
@@ -7,7 +7,7 @@ using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 
-namespace Microsoft.Azure.Templates.Analyzer.JsonRuleEngine.UnitTests
+namespace Microsoft.Azure.Templates.Analyzer.Utilities
 {
     [TestClass]
     public class JsonPathResolverTests
