+++ src/Analyzer.Utilities/Analyzer.Utilities.csproj
@@ -15,4 +15,8 @@
     <PackageReference Include="Newtonsoft.Json" Version="12.0.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\Analyzer.Types\Analyzer.Types.csproj" />
+  </ItemGroup>
+
 </Project>
